requirements.txt
Django==3.1.7
coverage==5.5
coveralls==3.0.1
django-cors-headers==3.7.0
djangorestframework==3.12.2
mysqlclient==2.0.3
django-test-without-migrations==0.6
web3==5.19.0